renamed Galaxy Window folder to GalaxyWindow
