+++ client/src/app/store.js
@@ -1,10 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
 import galaxyOptionsReducer from '../components/CreateGalaxy/galaxyOptionsSlice';
-import denseGalaxyPlanetSelectionReducer from '../components/Galaxy Window/denseGalaxySlice';
+import denseGalaxyPlanetSelectionReducer from '../components/GalaxyWindow/denseGalaxySlice';
 import userShipsReducer  from '../components/buildShips/UserShipSlice';
-import currentGalaxyIDReducer from '../components/Galaxy Window/galaxyWindowSlice';
-import missionQueueReducer from '../components/Galaxy Window/actionsToolbar/missionModule/missionModuleSlice';
-import toggleEndTurnReducer from '../components/Galaxy Window/menuSideSlice';
+import currentGalaxyIDReducer from '../components/GalaxyWindow/galaxyWindowSlice';
+import missionQueueReducer from '../components/GalaxyWindow/actionsToolbar/missionModule/missionModuleSlice';
+import toggleEndTurnReducer from '../components/GalaxyWindow/menuSideSlice';
 import currencyUpdateFlagReducer from '../components/TaskTracker/currencyUpdateFlag.js';
 
 export default configureStore({
