+++ client/src/components/GalaxyWindow/Chats/MainDisplay.jsx
