+++ client/src/components/GalaxyWindow/DenseGalaxy.jsx
