+++ client/src/components/GalaxyWindow/GalaxyWindow.jsx
