+++ client/src/components/GalaxyWindow/MenuSide.jsx
