+++ client/src/components/GalaxyWindow/SparseGalaxy.jsx
