+++ client/src/components/GalaxyWindow/actionsToolbar/missionModule/ShipListEntry.jsx
