+++ client/src/components/GalaxyWindow/actionsToolbar/missionSequence/MissionResult.jsx
