+++ client/src/components/GalaxyWindow/actionsToolbar/missionSequence/MissionResultTableRow.jsx
