+++ client/src/components/GalaxyWindow/actionsToolbar/missionSequence/missionType/Colony.jsx
