+++ client/src/components/GalaxyWindow/actionsToolbar/missionSequence/missionType/Scout.jsx
