+++ client/src/components/GalaxyWindow/denseGalaxySlice.js
