+++ client/src/components/GalaxyWindow/densePlanets.js
