+++ client/src/components/GalaxyWindow/galaxyWindowSlice.js
