+++ client/src/components/GalaxyWindow/sparseGalaxySlice.js
