+++ client/src/components/GalaxyWindow/sparsePlanets.js
